+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.16 branch
 ### 2.16.5-RC3
+* improved the `InstallShell` class;
 * Css files and KCFinder files to be loaded into the CKEditor editor are
     automatically set;
 * setted some default values for tables and iframe for CKEditor.